리뷰 + 별점 기능 개선

+++ src/main/java/com/geonpil/domain/Review.java
@@ -11,7 +11,7 @@ public class Review {
     private Long bookId;
     private Long userId;
     private String nickname;       // 사용자 닉네임 (조회 시용)
-    private Integer rating;        // 1~5점
+    private Double rating;        // 1~5점
     private String content;
     private Boolean isDeleted;
     private LocalDateTime createdAt;
